separate onclick and change tabs events
